added release date and Readme/Changelog/Bugs links, re-added WR842
Without the date, it was completely unclear to the end user whether
the page is current.
As the README and CHANGELOG are not part of the download, there was
almost no way end users could possibly find these.
We have a bug list so there's no reason to hide it.
The WR842 is an 8MB Flash device - re-included it in the download
list.